Only show current enrollments in Grades widget
refs: MBL-13149
affects: student
release note: Fixed a bug that could cause assignments for concluded enrollments to show in the Grades widget
